requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astroid==3.2.1
attrs==23.2.0
dill==0.3.8
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
isort==5.13.2
mccabe==0.7.0
multidict==6.0.5
pillow==10.3.0
platformdirs==4.2.2
pylint==3.2.0
python-dotenv==1.0.1
tomlkit==0.12.5
yarl==1.9.4
